+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-08-31
+### Added
+- Add Moonriver Network. Available under `MOVR` or `moonriver` ids.
+- Add option to change user-agent. `--user` or `-u` CLI option. Default user agent is `polkadot-rewards/version`
+
+### Fixed
+- Fix subscan URL. Used to be `network.subscan.io`, new url is `network.api.subscan.io`
+- Fix hitting API Request limit. Change subscan requests to 3 per second.
+
 ## [0.1.0] - 2021-08-08
 ### Added
 - Polkadot Rewards CLI that scrapes rewards for Polkadot & Kusama from Subscan & Coingecko